ch: rootfs: Chroot may exit none zero, keep track of them
This patch will track the shell runs on qemu, after this commit, the ci
could track the error in the qemu state and report ci failed.

+++ scripts/gen_image.sh
@@ -82,6 +82,12 @@ build_adjust() {
 	# Change to new rootfs and run the script
 	chroot ${BUILD_DIR}/rootfs /do_adjust.sh
 
+	if [ $? -ne 0 ]
+	then
+		echo "Some thing go wrong in chroot!"
+		exit 1
+	fi
+
 	# When the script is return, now it is in host env, remove the script
 	rm ${BUILD_DIR}/rootfs/do_adjust.sh
